Add type annotations and some Docblocks with linter hints
some minor code adjustments due to linter hints

+++ init.php
@@ -1,8 +1,8 @@
 <?php
 
 class ff_FeedCleaner extends Plugin {
-	private $host;
-	private $debug;
+	/** @psalm-suppress PropertyNotSetInConstructor */
+	private PluginHost $host;
 
 	function about() {
 		return array(
@@ -21,7 +21,7 @@ class ff_FeedCleaner extends Plugin {
 		return true;
 	}
 
-
+	/** @psalm-suppress MixedArgument */
 	function init($host) {
 		$this->host = $host;
 
@@ -31,12 +31,19 @@ class ff_FeedCleaner extends Plugin {
 	}
 
 
-	protected static function escape($msg) {
+	protected static function escape(string $msg): string {
 		return htmlspecialchars($msg, ENT_NOQUOTES);
 	}
 
-
-	private static function debug($msg, $prio=NULL) {
+	/**
+	 * Wrapper for debug messages.
+	 *
+	 * @param string $msg
+	 * @psalm-param 256|512|1024|null $prio
+	 * @return void
+	 * @see https://www.php.net/manual/errorfunc.constants.php
+	 */
+	private static function debug(string $msg, ?int $prio=null): void {
 		if(! class_exists("Debug")) {
 			trigger_error("Debug class doesn't exist. Why?", E_USER_ERROR);
 			return;
@@ -45,7 +52,7 @@ class ff_FeedCleaner extends Plugin {
 
 		Debug::log($msg, Debug::LOG_VERBOSE);
 		// TODO maybe try to detect the log destination to choose the necessary escaping?
-		if(is_int($prio)) trigger_error($msg, $prio);
+		if($prio) trigger_error($msg, $prio);
 	}
 
 
@@ -62,8 +69,18 @@ class ff_FeedCleaner extends Plugin {
 		return $feed_data;
 	}
 
-
-	private static function hook1($feed_data, $fetch_url, $json_conf) {
+	/**
+	 * Match the stored configurations to *$fetch_url*, execute textual ones right away and return
+	 * XML mods for later use.
+	 *
+	 * @param string $feed_data
+	 * @param string $fetch_url
+	 * @param string $json_conf
+	 * @return array{0: string, 1: array} first one is eventually modified feed data, second one is array of configs
+	 * @throws RuntimeException
+	 */
+	private static function hook1(string $feed_data, string $fetch_url, string $json_conf): array {
+		/** @psalm-suppress MixedAssignment */
 		$json_conf = json_decode($json_conf, true);
 		if (! $json_conf || ! is_array($json_conf)) {
 			throw new RuntimeException('No or malformed configuration stored. Possible cause: '. json_last_error_msg());
@@ -71,14 +88,14 @@ class ff_FeedCleaner extends Plugin {
 
 		$later = array();
 		foreach($json_conf as $config) {
-			$test = false;
+			if(! is_array($config)) continue;
 
-			if(array_key_exists('URL', $config))
-				$test = (strpos($fetch_url, $config['URL']) !== false);
-			elseif(array_key_exists('URL_re', $config))
-				$test = (preg_match($config['URL_re'], $fetch_url) === 1);
+			$url_ = $config['URL'] ?? null;
+			$url_re = $config['URL_re'] ?? null;
+			if(is_string($url_)) $test = (stripos($fetch_url, $url_) !== false);
+			elseif(is_string($url_re)) $test = (preg_match($url_re, $fetch_url) === 1);
 			else {
-				$msg = sprintf("Config '%s': Neither URL nor URL_re key is present", json_encode($config));
+				$msg = sprintf("Config '%s': Neither URL nor URL_re key is present, or wrong type", json_encode($config));
 				self::debug($msg, E_USER_WARNING);
 				continue;
 			}
@@ -86,7 +103,7 @@ class ff_FeedCleaner extends Plugin {
 			if( ! $test ) continue;
 
 			$msg = "Modifying '$fetch_url' using " . json_encode($config);
-			switch (strtolower($config["type"])) {
+			switch (strtolower($config["type"] ?? '')) {
 			case "regex":
 				self::debug($msg);
 				$feed_data = self::apply_regex($feed_data, $config);
@@ -104,7 +121,7 @@ class ff_FeedCleaner extends Plugin {
 			}
 		}
 
-		return array($feed_data, $later);
+		return [$feed_data, $later];
 	}
 
 
@@ -114,13 +131,14 @@ class ff_FeedCleaner extends Plugin {
 	}
 
 
-	private static function hook2($rss, $config_data) {
+	private static function hook2(FeedParser $rss, array $config_data): void {
 		static $p_xpath;
 		if(!$p_xpath) { #initialize reflection
-			$ref = new ReflectionClass('FeedParser');
+			$ref = new ReflectionClass(FeedParser::class);  # TODO use variable
 			$p_xpath = $ref->getProperty('xpath');
 			$p_xpath->setAccessible(true);
 		}
+		/** @var DOMXPath */
 		$xpath = $p_xpath->getValue($rss);
 		//$xpath->registerNamespace('rssfake', "http://purl.org/rss/1.0/");
 
@@ -140,22 +158,17 @@ class ff_FeedCleaner extends Plugin {
 	}
 
 
-	private static function link_regex($rss, $config) {
-		static $p_elem;
-		if(!$p_elem) {
-			$ref = new ReflectionClass('FeedItem_Common');
-			$p_elem = $ref->getProperty('elem');
-			$p_elem->setAccessible(true);
-		}
+	private static function link_regex(FeedParser $rss, array $config): void {
 		//  TODO link may be relative
 		//  URLHelper::rewrite_relative($site_url, $item->get_link());
 
 		$counter = 0;
 		foreach($rss->get_items() as $item_caps) {
 			$url = $item_caps->get_link();
-			$new_url = self::apply_regex($url, $config, True);
+			$new_url = self::apply_regex($url, $config, true);
 			if($new_url !== $url) {
-				$item = $p_elem->getValue($item_caps);
+				$item = $item_caps->get_element();
+				/** @psalm-suppress PossiblyNullReference */
 				$link = $item->ownerDocument->createElementNS("http://www.w3.org/2005/Atom", 'link');
 				$link->setAttribute('href', $new_url);
 				$item->insertBefore($link, $item->firstChild);
@@ -173,7 +186,7 @@ class ff_FeedCleaner extends Plugin {
 	}
 
 
-	private static function enc_utf8($feed_data, $config) {
+	private static function enc_utf8(string $feed_data, array $config): string {
 		$decl_regex =
 			'/^(<\?xml
 				[\t\n\r ]+version[\t\n\r ]*=[\t\n\r ]*["\']1\.[0-9]+["\']
@@ -196,13 +209,13 @@ class ff_FeedCleaner extends Plugin {
 	}
 
 
-	private static function apply_regex($feed_data, $config, $silent=False) {
+	private static function apply_regex(string $feed_data, array $config, bool $silent=false): string {
 		$pat = $config["pattern"];
 		$rep = $config["replacement"];
 
 		$feed_data_mod = preg_replace($pat, $rep, $feed_data, -1, $count);
 
-		if($feed_data_mod !== NULL) {
+		if($feed_data_mod !== null) {
 			$feed_data = $feed_data_mod;
 			if(! $silent) self::debug("Applied (pattern '$pat', replacement '$rep') $count times");
 		} else {
@@ -213,11 +226,10 @@ class ff_FeedCleaner extends Plugin {
 	}
 
 
-	private static function apply_xpath_regex($xpath, $config) {
-		if(isset($config['namespaces']) && is_array($config['namespaces']))
-			foreach($config['namespaces'] as $prefix => $URI)
-				$xpath->registerNamespace($prefix, $URI);
-		else { //TODO remove this
+	private static function apply_xpath_regex(DOMXPath $xpath, array $config): void {
+		if(isset($config['namespaces']) && is_array($config['namespaces'])) {
+			foreach($config['namespaces'] as $prefix => $URI) $xpath->registerNamespace($prefix, $URI);
+		} else { //TODO remove this
 			$DNS = array(
 			"http://www.w3.org/2005/Atom",
 			"http://purl.org/rss/1.0/",
@@ -238,10 +250,10 @@ class ff_FeedCleaner extends Plugin {
 
 		self::debug("Found {$node_list->length} nodes with XPath '{$config['xpath']}'");
 
-		$preg_rep_func = function($node) use ($pat, $rep, &$counter) {
+		$preg_rep_func = function(DOMNode $node) use ($pat, $rep, &$counter): void {
 			if( $node->nodeType == XML_TEXT_NODE) {
 				$text_mod = preg_replace($pat, $rep, $node->textContent, -1, $count);
-				if($text_mod !== NULL) {
+				if($text_mod !== null) {
 					$node->nodeValue = $text_mod;
 					$counter += $count;
 				} else {
@@ -249,6 +261,7 @@ class ff_FeedCleaner extends Plugin {
 				}
 			}
 		};
+
 		$counter = 0;
 		foreach($node_list as $node) {
 			$preg_rep_func($node);
@@ -278,7 +291,7 @@ class ff_FeedCleaner extends Plugin {
 <?php
 	}
 
-	function index() {
+	public function index(): void {
 		$pluginhost = PluginHost::getInstance();
 		$json_conf = $pluginhost->get($this, 'json_conf');
 
@@ -323,12 +336,12 @@ class ff_FeedCleaner extends Plugin {
 <?php
 	}
 
-	function save() {
-		$json_conf = $_POST['json_conf'];
+	public function save(): void {
+		$json_conf = $_POST['json_conf'] ?? null;
 
 		if(is_null(json_decode($json_conf))) {
 			echo json_encode(["errMsg" => __("Invalid JSON. Possible Reason: ") . json_last_error_msg()]);
-			return false;
+			return;
 		}
 
 		$this->host->set($this, 'json_conf', $json_conf);
@@ -338,17 +351,22 @@ class ff_FeedCleaner extends Plugin {
 
 	// diff stuff
 
-	static function format_diff_array_html($ar) {
+	/**
+	 * Formats an array of strings into presentable HTML. Each entry becomes a line, seperated by <br>s.
+	 *
+	 * @param string[] $ar
+	 * @return string
+	 */
+	private static function format_diff_array_html(array $ar): string {
 		// TODO should make sure that everything is indeed utf-8.
 		// This may be not the case if below, the feed data is not loaded into a xml doc.
 		// (and not even then?)
-		$func = function($var) {return htmlspecialchars($var, ENT_QUOTES, "UTF-8");};//ENT_XML1 would be better?
-		$diff = array_map($func, $ar);
+		$diff = array_map(fn (string $var) => htmlspecialchars($var, ENT_QUOTES, "UTF-8"), $ar); # TODO ENT_XML1 would be better?
 
 		return implode("<br/>", $diff);
 	}
 
-	function preview() {
+	public function preview(): void {
 		$url = $_POST['url'];
 		$conf = $_POST['json_conf'];
 
@@ -360,36 +378,49 @@ class ff_FeedCleaner extends Plugin {
 		}
 	}
 
-	static function format_save_tmp($data, $format=True) {
+	/**
+	 * Stores given XML data to a temp file and returns its file name. Formats the XML nicely by default.
+	 *
+	 * @psalm-param non-empty-string $data
+	 * @param boolean $format
+	 * @return string
+	 * @throws RuntimeException
+	 */
+	private static function format_save_tmp(string $data, bool $format=true): string {
+		$filename = tempnam(sys_get_temp_dir(), '');
+		$xml = new DOMDocument();
+
 		if($format) {
-			$xml = new DOMDocument();
 			$xml->preserveWhiteSpace = false;
 			$xml->formatOutput = true;
-			$res = $xml->loadXML($data);
 		}
+		$res = $xml->loadXML($data);
+		$res2 = $xml->save($filename);
 
-		$filename = tempnam(sys_get_temp_dir(), '');
-		if($format && $res) $outXML = $xml->save($filename);
-		else {
-			$handle = fopen($filename, "w");
-			fwrite($handle, $data);
-			fclose($handle);
-		}
+		if(! $res || ! $res2) throw new RuntimeException("Coundn't write XML.");
 
-		return array($filename, $format && $res);
+		return $filename;
 	}
 
 	const diff_cmd = 'diff -U 0 -s -w ';
 
-	static function compute_diff($url, $json_data) {
+	/**
+	 * Compute a diff between original feed data found under *$url*,
+	 * and the feed after applying all modifications this plugin performs.
+	 *
+	 * @param string $url
+	 * @param string $json_data
+	 * @return array
+	 * @throws RuntimeException
+	 */
+	private static function compute_diff(string $url, string $json_data): array {
 		//TODO maybe use a library for computing the diff,
 		/* see http://stackoverflow.com/questions/321294/ or https://github.com/chrisboulton/php-diff
 		or https://github.com/sebastianbergmann/diff
 		*/
 		$con = UrlHelper::fetch($url);
-		if(!$con) throw new RuntimeException("Couldn't fetch '$url'");
+		if(!$con) throw new RuntimeException("Couldn't fetch '$url': " . UrlHelper::$fetch_last_error);
 
-		// could throw Exception
 		list($feed_data, $config_data) = self::hook1($con, $url, $json_data);
 
 		$rss = new FeedParser($feed_data);
@@ -399,16 +430,19 @@ class ff_FeedCleaner extends Plugin {
 
 		self::hook2($rss, $config_data);
 
-		$ref = new ReflectionClass('FeedParser');
+		$ref = new ReflectionClass(FeedParser::class);  # TODO use variable
 		$p_doc = $ref->getProperty('doc');
 		$p_doc->setAccessible(true);
 
+		/** @var DOMDocument */
 		$doc = $p_doc->getValue($rss);
 		$new_feed_data = $doc->saveXML();
 
-		list($old_file, $xml) = self::format_save_tmp($con);
-		list($new_file, $xml) = self::format_save_tmp($new_feed_data, $xml);
-		$diff = array();
+		if(! $new_feed_data) throw new RuntimeException("Error computing diff: modified feed is empty");
+
+		$old_file = self::format_save_tmp($con);
+		$new_file = self::format_save_tmp($new_feed_data);
+		$diff = [];
 		$res = 2;
 		exec(self::diff_cmd . " $old_file $new_file", $diff, $res);
 		unlink($old_file);
